\name{genotypes}
\docType{data}
\alias{genotypes}
\title{Example Genotype Data Set}
\description{
An example genotype matrix for the \pkg{hsphase} package.
}
\usage{
data(genotypes)
}
\format{
A genotype \code{matrix} with:
\itemize{
  \item \strong{Columns:} SNPs
  \item \strong{Rows:} animals/individuals
}
Genotype coding follows the package conventions (typically \code{0, 1, 2} and \code{9} for missing).
}
\references{Sahoo S., Ferdosi M.H., van der Werf J.H.J., and de las Heras-Saldana S., et al. (2025) \emph{Proc. Assoc. Advmt. Anim. Breed. Genet}. 26: 323}
\keyword{datasets}
