\name{map}
\docType{data}
\alias{map}
\title{Example Map File}
\description{
An example map dataset for the \pkg{hsphase} package.
}
\usage{
data(map)
}
\format{
A \code{data.frame} with the following columns:
\describe{
  \item{Name}{SNP identifier}
  \item{Chr}{Chromosome}
  \item{Position}{SNP position in base pairs}
}
}
\references{Sahoo S., Ferdosi M.H., van der Werf J.H.J., and de las Heras-Saldana S., et al. (2025) \emph{Proc. Assoc. Advmt. Anim. Breed. Genet}. 26: 323}
\keyword{datasets}
