\name{pedigree}
\docType{data}
\alias{pedigree}
\title{Example Pedigree}
\description{
An example pedigree dataset for the \pkg{hsphase} package.
}
\usage{
data(pedigree)
}
\format{
A \code{data.frame} with the following columns:
\describe{
  \item{First column}{Half-sib (offspring) IDs}
  \item{Second column}{Sire IDs}
}
}
\references{Sahoo S., Ferdosi M.H., van der Werf J.H.J., and de las Heras-Saldana S., et al. (2025) \emph{Proc. Assoc. Advmt. Anim. Breed. Genet}. 26: 323}
\keyword{datasets}
