\name{genotypes}
\docType{data}
\alias{genotypes}
\title{Example of Genotype Data Set}
\description{An example of genotype matrix for hsphase.  
}
\usage{
data(genotypes)
}
\format{a genotype matrix with the following columns and rows:
  \itemize{
    \item{Columns: }{SNPs}
    \item{Rows: }{Animals}
    }
}
\keyword{datasets}
