\name{genotypes}
\docType{data}
\alias{genotypes}
\title{Example of Genotype Data Set}
\description{
  This data set serves as an example of a genotype matrix intended for use with the \code{hsphase} package. 
  }
\usage{
data(genotypes)
}
\format{
  The data set is a genotype matrix with specific structure, including:
  \itemize{
    \item \strong{Columns:} Represent Single Nucleotide Polymorphisms (SNPs). Each column corresponds to a specific SNP.
    \item \strong{Rows:} Represent individual animals. Each row corresponds to the genotypic data for a single animal across various SNPs.
  }
}
\keyword{datasets}
