\name{map}
\docType{data}
\alias{map}
\title{Example of Map File}
\description{An example of map for hsphase.   
}
\usage{
data(map)
}
\format{a data frame with the following columns:
  \itemize{
    \item{Name: }{SNP name}
    \item{Chr: }{chromosome}
    \item{Position: }{SNP position in base pairs}
    }
}
\keyword{datasets}