\name{map}
\docType{data}
\alias{map}
\title{Example Map File for Genetic Data}
\description{
  This data set is an example of a map file used within the \code{hsphase} package to demonstrate the mapping of SNPs to their respective locations on chromosomes. 
}
\usage{
data(map)
}
\format{
  The data set is formatted as a data frame with the following columns, providing essential information about each SNP:
  \itemize{
    \item \strong{Name:} The unique identifier or name of the SNP.
    \item \strong{Chr:} The chromosome on which the SNP is located.
    \item \strong{Position:} The position of the SNP on the chromosome, expressed in base pairs.
  }
}
\keyword{datasets}

