\name{pedigree}
\docType{data}
\alias{pedigree}
\title{Example Pedigree}
\description{An example pedigree for hsphase.   
}
\usage{
data(pedigree)
}
\format{a data frame with the following columns:
  \itemize{
    \item{First Column: }{half-sibs}
    \item{Second Column: }{sires}

    }
}

\keyword{datasets}