\name{pedigree}
\docType{data}
\alias{pedigree}
\title{Example Pedigree Data Set}
\description{
  This dataset provides an example of a pedigree, specifically designed for use with the \code{hsphase} package.
}
\usage{
data(pedigree)
}
\format{
  The dataset is structured as a data frame with detailed familial relationships, including:
  \itemize{
    \item \strong{First Column:} Identifiers for half-sibs. 
    \item \strong{Second Column:} Identifiers for sires. 
  }
}
\keyword{datasets}

